app/requirements.txt
aiohttp==3.8.6
django-admin-interface==0.26.1
django-cors-headers==4.3.0
django-stubs-ext==4.2.5
django-types==0.19.1
djangorestframework==3.14.0
docopt==0.6.2
gql==3.4.1
pip-chill==1.0.3
psycopg==3.1.12
psycopg-binary==3.1.12
yarg==0.1.9
# aiosignal==1.3.1 # Installed as dependency for aiohttp
# asgiref==3.7.2 # Installed as dependency for django
# async-timeout==4.0.3 # Installed as dependency for aiohttp
# attrs==23.1.0 # Installed as dependency for aiohttp
# backoff==2.2.1 # Installed as dependency for gql
# certifi==2023.7.22 # Installed as dependency for requests
# charset-normalizer==3.3.1 # Installed as dependency for aiohttp, requests
# django==4.2.6 # Installed as dependency for django-cors-headers, django-stubs-ext, djangorestframework
# django-colorfield==0.10.1 # Installed as dependency for django-admin-interface
# frozenlist==1.4.0 # Installed as dependency for aiohttp, aiosignal
# graphql-core==3.2.3 # Installed as dependency for gql
# idna==3.4 # Installed as dependency for requests, yarl
# multidict==6.0.4 # Installed as dependency for aiohttp, yarl
# pillow==10.1.0 # Installed as dependency for django-colorfield
# python-slugify==8.0.1 # Installed as dependency for django-admin-interface
# pytz==2023.3.post1 # Installed as dependency for djangorestframework
# requests==2.31.0 # Installed as dependency for yarg
# sqlparse==0.4.4 # Installed as dependency for django
# text-unidecode==1.3 # Installed as dependency for python-slugify
# types-psycopg2==2.9.21.15 # Installed as dependency for django-types
# typing-extensions==4.8.0 # Installed as dependency for django-stubs-ext, psycopg
# urllib3==2.0.7 # Installed as dependency for requests
# yarl==1.9.2 # Installed as dependency for aiohttp, gql
